function that executes a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,5 @@
 #include "function_pointers.h"
+#include <stdlib.h>
 /**
  * array_iterator - executes a function pointer given as a parameter
  * on each element of an array
@@ -10,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-unsigned int i;
+size_t i;
 
 if (array && action)
 {
